expenses_bot/requirements.txt
python-telegram-bot==21.6
google-api-python-client==2.155.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
anthropic>=0.40.0
python-dotenv>=1.0.0
